Fix crash from /setblock or /fill with ~ syntax
`Integer.valueOf(match.group(...))` was crashing due to
NumberFormatException when trying to parse "~" syntax in /setblock or
/fill commands. Check for "~" syntax and print a warning to the user
that ~ syntax cannot be undone.

(A nicer fix would be to parse the ~ syntax and translate it to player
coordinates. But that could result in misalignment between the native ~
syntax and how Minescript interprets it.)