Fix onClientChatReceived handler on Fabric & Forge
Stop using the ClientChatReceivedEvent on Forge and implement an
equivalent mixin for Forge (ChatComponentMixin) and Fabric
(ChatHudMixin). Implementing this event handler as a mixin allows the
Forge and Fabric mods to have identical behavior when handling chat
events.

This fixes the script function "get_client_chat_received_events" for
both Fabric and Forge.

The onClientChatReceived was previously broken for the Forge mod and not
implemented for Fabric.

Remove the M1CocoaMixin which is no longer needed.