Update project version to 2.0
The previous version was 1.19.2. This changes the numbering scheme to
diverge from Minecraft version numbers. This establishes semantic
versioning for Minescript that makes it clear when there's a major vs
minor update to the mod. Also, it solves the problem of how to specify
versions when there are multiple Minescript updates, like bug fixes, for
a single Minecraft version.