Add script function flush() to minescript.py
minescript.flush() allows scripts to wait for all pending commands
issued by the script to complete. This is useful when it's important
to synchronize the script execution with the execution of the commands
output by the script.

E.g. the following script would print almost exactly the same time
throughout the loop as the commands are buffered faster than Minecraft
can process them:

for i in range(1000):
  minescript.exec(f"setblock {x} {y} {z} yellow_concrete")
  minescript.chat(time.time())

With the introduction of flush(), the timestamps more accurately reflect
when the setblock commands are executed:

for i in range(1000):
  minescript.exec(f"setblock {x} {y} {z} yellow_concrete")
  minescript.flush()
  minescript.chat(time.time())

Note that flush() can slow script execution considerably because it
requires (potentially several) roundtrips of the tick cycle.