Change ChunkLoadEventListener to take Runnable
Previously, ChunkLoadEventListener took ScriptFunction which it would
use to respond when all requested chunks have been loaded. Generalize
that param to Runnable so that ChunkLoadEventListener can be used in
other situations unrelated to script functions, e.g. allowing the
getblock script function to optionally wait for the requested block to
be loaded before reading block state.