Overhaul copy/paste commands with smarter limits
Eliminate minescript_copy_size_limit which defaulted to 200 blocks in
each direction. Replace it with a soft limit of 1600 chunks, which can
be overridden by passing `no_limit` to copy and paste. Also, copy and
paste commands now automatically detect if any chunks within the
copied/pasted region are unloaded, and the commands will exit
immediately with an error if that's the case. This significantly reduces
the chances of incomplete copies/pastes (however, the user can still
break a copy/paste operation by moving or teleport to unload chunks in
the middle of the operation).

Added getblocklist(...) which is a batch form of getblock(...) which
takes a list of (x, y, z) tuples/lists and returns a corresponding set
of blocks at those positions. getblocklist(...) is needed to optimize
performance of the new paste.py implementation, as calling getblock(...)
hundreds of times to detect whether chunks are loaded in the paste
region is prohibitively expensive. Hundreds of getblock(...) calls could
take a minute, whereas a single getblocklist(...) call for the same
number of blocks completes in a fraction of a second.