Add version-check flags to minescript_runtime.py
These flags can be added to any Python script that imports minescript or
minescript_runtime:

  --debug-version-check: print debug info about version checking
  --relax-version-check: don't exit upon version errors, run anyway
  --version-check-only: exit immediately after performing version checks

These flags are stripped from sys.argv so that scripts that import
minescript or minescript_runtime do not see them.

Note that version requirements are always checked regardless of which,
if any, of these flags are given.