Initial impl of Volume (to be renamed BlockPack)
Volume (soon to be renamed "BlockPacker" and "BlockPack") is a way of
building a compact representation of a volume of blocks. This is
intended initially for use with the copy, paste, and undo commands so
that the sets of blocks that are written to files are more compact.

Currently, copy and undo files contain only setblock commands. This is
wasteful because often there are large adjacent sections of the same
type of block which could be much more efficiently represented by a
fill command. Volume/BlockPacker identifies embedded fill volumes of the
same type of block within an arbitrary set of blocks.