Reuse block type IDs for overwritten unique blocks
Previously, block type IDs within a BlockPacker.Tile could exhaust 15
bits when overwriting blocks within the tile, thereby running out of
IDs. Track the overwriting of blocks so that when the last block of a
specific type in a tile is overwritten, that ID is added to a freelist
to be used for allocating new block type IDs.