Implement BlockPack writeFile and readFile
Initial implementation of BlockPack file IO as an uncompressed zip file;
uncompressed because it's a binary format that doesn't compress well
(aside from the symbol table of block types), and zip format so it's
easily transportable across systems and networks.

Also, refactor BlockPack.Tile.getCoord(...) to pass a short value rather
than a short[] and index into it.