Implement undo command using BlockPack
Compared to the prior implementation, the BlockPack implementation is
about 10-15x faster, and 80-150x more space efficient (both in memory
and on disk).

Using BlockPack for undo commands is enabled by default, and can be
disabled by running the Minescript command:

  \minescript_use_blockpack_for_undo false

or via setting in minescript/config.txt:

  minescript_use_blockpack_for_undo=false

If disabled, the prior implementation is used, storing setblock commands
one per line in a text file.