Support rotations in BlockPacker.add_blockpack()
Support `rotation` param in addition to `offset`. Note that the rotation
is applied to the BlockPack coordinates before the offset. Rotation is
representation as a flattened 3x3 row-major matrix of ints, with a
default of the identity rotation matrix: (1, 0, 0, 0, 1, 0, 0, 0, 1)

Or in matrix format:
  (1, 0, 0,  # coefficients for resulting x component
   0, 1, 0,  # coefficients for resulting y component
   0, 0, 1)  # coefficients for resulting z component

E.g. to flip the z coordinate use the rotation:

  (1, 0, 0, 0, 1, 0, 0, 0, -1)

or to swap x and z:

  (0, 0, 1, 0, 1, 0, 1, 0, 0)

Note that blocks that face a specific direction are currently not
rotated to face in a different direction, only the coordinates are
rotated. There is a TODO to fix this.