Improve handling of blockpack file IO errors
Handle blockpack file IO errors differently in distinct cases. In one
case, the user has run a script that has request reading or writing a
file that doesn't exist. These errors are reported in the in-game HUD as
exception.getMessage() from Java. The other case is when a built-in
Minescript command like `undo` fails to read or write a blockpack file,
in which case a more verbose error is printed to the in-game HUD
including the Java stacktrace to facilitate debugging of Minescript
itself.