Implement autorun to run commands when join world
Summary of changes:
- new `autorun[world]=\command` configuration in config.txt
- script jobs are now automatically killed when exiting a world
- remove support for scripts to persist at world exit with on_world_exit

Details...

Support lines on config.txt that specify "autorun" commands. The form of
these config lines is:

autorun[World Name]=\command_to_run with args

The special name `*` indicates that the command should be run when
entering all worlds:

autorun[*]=\command_to_run with args

The command to the right of the first `=` in the autorun config line is
interpreted just as if entered by hand into the chat. Multiple config
lines can be configured for the same world or for `*`, in which multiple
commands are run concurrently when entering matching worlds.

Also, replace "server_name" and "server_address" attributes returned
from minescript.world_properties() with "name" and "address" so that
these two attributes are populated for both server worlds and local
worlds. (Address is "localhost" for local worlds.)

Remove support for scripts to persist when leaving a world and getting
notified of entering/exiting worlds (on_world_enter(...) and
on_world_exit(...) in minescript.py). This feature was hard to use
because it would break the event loop in minescript_runtime.py when
calling script functions with return values.