Fix timing bug with autorun when entering world
Timing bug would happen when the thread that listens for the player
entering/exiting worlds detected that a world was entered, but later
than onPlayerTick() was called. This would result in the autorun
commands being run twice.

The fix is to call `autorunHandled.set(false)` only when a world is
exited, and in onPlayerTick() use autorunHandled.getAndSet(true)
instead of separate autorunHandled.get() and autorunHandled.set(true)
which are collectively not atomic.

Also, call systemCommandQueue.clear() in onPlayerTick() when
autorunHandled is false, but not when a world enter event is detected on
the world listener thread.