Support multi-line config vars in config.txt
Lines ending with a single backslash are concatenated with the following
line. This is convenient for configuring multiple semicolon-delimited
Minescript commands to run in sequence when entering a world. E.g.

autorun[*]=\
    eval '"*** Welcome to %s ***" % world_properties()["name"]'; \
    eval world_properties()

Also, treat autorun commands that lack a slash or backslash prefix as
Minescript commands by implicitly prepending a backslash.