Use same source mappings for Forge and Fabric mods
Change Fabric mod to officialMojangMappings so that it uses the same
mappings as the Forge mod. This allows the common Minescript.java to be
identical across the two mods, eliminating the need for remapping of
symbols via rename_minecraft_symbols.py.

Introduce a Platform interface to differentiate the two different
obfuscated names for the ChatScreen.input field which is accessed by
name via reflection.