Call Minescript.init() on Forge only client-side
Previously, Minescript.init(...) was called unconditionally from the
MinescriptForgeMod constructor, even on Forge servers. This had the side
effect of creating a minescript directory and copying various Python
scripts to it. Now, init() is called only when initializing the Forge
mod on a client, matching the behavior of the Fabric mod.