Return Java exception info, stacktrace to Python
When Java code running within handleScriptFunction(...) throws an
exception, translate it to an ExceptionInfo record which contains the
stacktrace up to and through Minescript.java. ExceptionInfo is
translated into JSON and returned to Python, where it's translated to
the JavaException dataclass and raised as a Python exception.

Previously, Java exceptions were translated on the Java side to Python
exception types. Now, a Java exception's type, message, and stacktrace
are preserved when translated to a Python exception.