Throw exceptions for bad script function calls
Throw exceptions from Java when script functions have the wrong number
or types of args, or illegal states (e.g. requesting an operation on a
BlockPack which does not exist). These exceptions are passed to Python
and raised as JavaException.

Previsouly, many cases of bad args or illegal states were handled by
logging the errors to the game chat.