Support scripts in subdirs, configure script path
Support execution of scripts from dirs specified in configurable script
path. E.g. in config.txt:

path=:foo:bar/baz:/Users/maxuser/scripts

The above example configures 4 dirs containing scripts that can be
tab-completed from the chat:

1. the empty dir, i.e. the `minescript` dir itself
2. `minescript/foo`
3. `minescript/bar/baz`
4. `/Users/maxuser/scripts` (absolute path)

Script names in subdirs can be tab-completed using names relative to
dirs in path.

Windows uses ";" as the path separator, and UNIX-like systems use ":".

The default value for `path` is effectively the `minescript` dir.

When executing scripts as subprocesses, set PYTHONPATH to the
`minescript` dir so that scripts run from other dirs can `import
minescript`.

Also, introduce `which` command to echo to the chat the fully qualified
dir containing the specified script, or indication that the command is a
built-in. E.g. `\which eval` or `\which ls`.