Support alternative script languages for commands
In addition to the default support for Python, support other languages
for commands. In config.txt, the following for UNIX-like systems:

  python="/usr/bin/python3"

is equivalent to the following:

  command = {
    "extension": ".py",
    "command": [ "/usr/bin/python3", "-u", "{command}", "{args}" ],
    "environment": [ "PYTHONPATH={minescript_dir}" ]
  }

When expanding the command to be executed, the filename extension (e.g.
".py") is automatically appended to "{command}", "{args}" is expanded to
the given args, and "{minescript_dir}" is expanded to the location of
the `minescript` dir. The value of the `command` variable is in JSON
format.

E.g. support for executing .jar files via `java` can be configured with:

  command = {
    "extension": ".jar",
    "command": [ "/usr/bin/java", "-jar", "{command}", "{args}" ]
  }

"environment" for specifying environment variables to the command
subprocess is optional.

Multiple `command` variables can be specified in config.txt for
supporting execution of multiple different file types.

Also, config.txt now supports multi-line variables where the value
starts with `{`, with the value terminating with the line that ends with
`}`.