Relocate built-in scripts to minescript/system
When loading Minescript Python scripts from the jar resources to the
`minescript` dir, place libraries (to be imported from other scripts) in
`minescript/system/lib` and executable scripts (runnable from chat) in
`minescript/system/exec`.

Automatically delete built-in scripts appearing in the legacy location,
i.e. directly in the `minescript` dir.

Remove the special case of filtering scripts with names starting with
"minescript" from tab completion, since minescript.py and
minescript_runtime.py are no longer tab-completed due to their new
location. Those scripts are now in `minescript/system/lib` which is in
PYTHONPATH (so they can be imported) but that dir is not in the command
path used for tab completing runnable commands (which effectively
defaults to `minescript/system/exec:minescript`).

Also, add built-in command `reload_minescript_resources` to force reload
scripts from the Minescript jar resources to `minescript/system` files.
This is helpful during development so that `version.txt` doesn't need to
be deleted to force resources to be reloaded when a new jar is built.