Escape double quotes in commands for Windows
Support escaping of double quotes in commands, which is needed for
command interpretation on Windows which does shell interpretation of
double quotes in the command array passed to `Runtime.exec(...)`.
Escaping is enabled by default for Windows, and can be enabled or
disable explicitly via config variable in config.txt, e.g.

```
escape_command_double_quotes=true
```

Add a unittest in minescript_test.py for escaping of quotes and double
quotes in commands: command_parse_test. This test can be run in
isolation as:

```
\minescript_test command_parse_test
```