Move version.txt to minescript/system dir
Previously, version.txt lived directly in the minescript dir. Moving it
to the minescript/system dir keeps it out of the way of user-edited
files directly in the minescript dir.

During Minescript init, if the legacy version.txt file is found, delete
obsolete files created by legacy versions directly in the minescript
dir.