Add show_chat_screen(show, prompt) script function
`show_chat_screen(...)`  allows scripts to show or hide the chat screen.
When showing the chat screen, an optional string prompt can be given to
be inserted into the edit box if the chat screen is being shown.

This function is useful for adding prompt prefixes into the chat for use
with a chat interceptor (see ChatMessageInterceptor). A chat interceptor
can filter messages so that only messages matching the given prompt
prefix are intercepted for custom handling while others messages are
sent to the chat or executed as commands.