Implement help for built-in commands with `\help`
If command requested by `help` is not a built-in, fall back to external
help.py script for reading pydoc of the requested command/script. Update
help.py to echo a message to chat when an external script is found but
has no pydoc, printing the source location of the script.

Also, fix error in `\config` due to typo when getting "debug_output".