Improve handling of JsonSyntaxException for args
Parsing script function args in the presence of unsynchronized output
from a script's stdout (using via print() calls) can result in malformed
JSON which triggers JsonSyntaxException. Raise a new exception
explaining the likely cause - and fix - so that it's handled at the call
site of the corresponding script function.

Previously, the original JsonSyntaxException was echoed to the chat and
logged, but not raised as a Python exception from the script function
that encoutered the error, and failed to terminate the script function,
hanging the enclosing job.