Improve usability of echo() script function
To echo a single arg that's a dict or list, but not interpret it as
JSON-formatted text, allow empty string as an additional arg to treat
the input as a string. E.g.

echo([1, 2, 3], "")

Previously, passing a dict or list as the first arg to `echo()` would
ignore subsequent args.