Fix file handle leak from Files.list()
Wrap `Files.list(resolvedDir)` from ScriptConfig in a try-with-resources
block so that the stream is closed, along with the file handles for the
directories managed by the stream.

Verified on macOS with the following command to list the most frequently
leaked file handles:

$ lsof -p "$(ps auxwww |grep minecraft |awk '{print $2}' |xargs |sed 's/ /,/g')" |awk '{print $9}' |sort |uniq -c |sort -k 1 -g |tail

File handle counts, in order of increasing frequency, after executing
several commands before the fix, using `path=system/exec::/:foo:bar/baz`
in `config.txt`:

...
  42 saves/TestWorld/...
1388 /
1388 minescript
1388 minescript/bar/baz
1388 minescript/foo
1388 minescript/system/exec

and after the fix:

...
  42 saves/TestWorld/...

Previously, file handles for directories in the command path (dirs
specified by `path` config variable in Config.java) were owned by a
Stream<Path> that was not explicitly closed. As a result, the stream
(and the directories associated with it) was not getting finalized in a
reasonable amount of time (possibly ever), effectively leaking the file
handles of those directories until exhausting the Java process's open
file-handle limit, after which the JVM would throw a subclass of
IOException when attempting to open additional files.