Update Java reflection script functions
Replace MethodSet with MemberSet and add support for Java fields. A
member has an optional field and a list of overloaded methods. This
allows Python code to refer to a member of a class by name and access it
as a field or method call.

Add `java_double(...)` for creating Java doubles.

Rename Java-reflection script functions to all use a "java_" prefix.