Make key events consistent across mod loaders
Move the fabric KeyboardMixin to the common KeyboardHandlerMixin so that
all mod loaders are using the same keyboard mixin for relaying key
events.

Previously, Forge and NeoForge were using loader-specific keyboard
handlers which passed along key events when the chat screen was absent,
but no key events when the chat screen was present.