Create branch for mc1.20.6-forge
MultiLoader-Template doesn't have support Forge with MC 1.20.6.  The MLT
configs for 1.20.4 and 1.21 use incompatible build-config structures
which are nontrivial to reconcile.

As a workaround, the Minescript mc1.20.6 branch continues to support
only Fabric and NeoForge (based on the official MLT template for that MC
version) while this new mc1.20.6-forge branch supports only Forge.