Allow Java interfaces to be implemented in Python
InterfaceProxy is a generalization of the former FunctionalParamProxy
which allowed script functions to be passed to Java methods that accept
functional types (i.e. those accepting lambdas). InterfaceProxy now
supports creating a proxy for any Java interface.

InterfaceProxy is instantiated automatically as an implementation detail
when a script function is passed to a Java method expecting a param
whose type is an interface. When the interface has exactly one abstract
method, the script function is passed the args of that method when the
method is invoked, e.g.

  def my_script_function(args_passed_to_java_method):
   ...

When the interface has multiple abstract methods, the script function is
passed the Java Method object that was invoked plus its args as
Object[]:

  def my_script_function(java_method, args_passed_to_java_method):
    ...

Also, add support for bound methods, i.e. Python objects bound to a
method for deferred invocation with args. E.g.

  object = SomePythonType()
  binding = object.foo  # requires that object has no field `foo`
  binding(args)  # calls object.foo(args)