Typing corrections (#29)
Added and fixed some parameters and return types.
Keep in mind that this code still uses the syntax prior to Python 3.9. Nowadays, the typing module is no longer used, collections.abc is used, "|" is used instead of Union, the built-in "list" is used instead of "List", etc.
It would be great to update it at some point.