Integrate Pyjinn scripts into job management
When running a Pyjinn script from in-game chat, create a script job to
manage it.  Event listeners are not yet supported in Pyjinn scripts, and
script jobs currently immediately close after executing all global
statements.

Refactored subprocess-specific aspects of Job into Job.ExternalJob and
made Job an abstract class so it can be subclassed for internally
executed Pyjinn scripts. Changed PyjinnScript.create() to
PyjinnScript.createJob() which so that it can be managed by JobManager.