Partial support for Pyjinn event listeners
Add initial support to Pyjinn scripts for the various
register_<event>_listener script functions, but with a different
function signature (taking a single callback param) from the Python API
in minescript.py (which takes no params, and requires a follow-up call
to start_<event>_listener passing the ID returned from the register
call)):

- register_key_listener(callback)
- register_mouse_listener(callback)
- register_chat_message_listener(callback)
- register_chat_message_interceptor(callback)
- register_add_entity_listener(callback)
- register_block_update_listener(callback)
- register_explosion_listener(callback)
- register_take_item_listener(callback)
- register_damage_listener(callback)
- register_chunk_listener(callback)

TODO: The registered events trigger PyjinnTask::sendResponse, but the
functionCallId is not yet associated with the callback function passed
from the Pyjinn script, so the callback isn't called yet.  Either the
script/job/task needs to maintain a mapping from functionCallId (long)
to callback (Script.Function) or the callback needs to be wired
directly throughout in place of the numeric functionCallId. Also,
killing a Pyjinn job currently fails to close the job and doesn't stop
the event from triggering Task::sendResponse.