Relax replacement of %userprofile% in Windows path
Substitute the user's home directory (`System.getProperty("user.home")`)
in place of "%userprofile%" anywhere that it appears in the `python`
config variable in `config.txt` on Windows.

Previously, "%userprofile%" in the `python` config variable on Windows
was required to be at the beginning of the string and followed by a
backslash.