Relocate jar resources to match output paths
Previously, all the jar resources were in the same directory, which
required the Pyjinn version of minescript.py to have a different name as
a jar resource (minescript.pyj) as it did when copied to the user's file
system. The renaming avoided a name conflict with the Python version of
mienscript.py. Now, the Pyjinn and Python lib files are stored in
different directories within the jar resources, eliminating the need to
rename files.