Fix bug in embedded Pyjinn scripts on Windows
The script name used for embedded Pyjinn scripts is not associated with
an actual file. But the ":" character used in the script name throws an
exception on Windows when interpreted as a filename via
java.nio.file.Path.

Fix PyjinnScript.loadScript() by not interpreting the script name with
Paths.get(scriptFilename).getFileName().