Fix missing JobInfo.self returned from job_info()
Always populate JobInfo.self, which is True only for the enclosing job,
with a boolean value and False otherwise.

Previously, the `self` attribute of JobInfo was omitted for jobs other
than the current job which has a self value of True.

Also, change the name of embedded Pyjinn jobs from the pattern
"eval_pyjinn_script.%d.%d" to "eval_pyjinn_script-%d-%d" for
improved readability in chat.