Support embedded Pyjinn script in "with" block
Example:

```
import java

with java.eval_pyjinn_script("x = 3") as script:
  print(script.get("x"))
```

When the "with" block exits, script.exit() is called automatically.

Introduce ScriptObject in java.py as a subclass of JavaObject for
supporting context manager semantics for "with" blocks. Move ownership
of script.script_env JavaHandle to ScriptObject so that JavaObject
itself never owns script_env.